requirements.txt
numpy
scipy
pyyaml
matplotlib
pandas
streamlit
